+++ app/controllers/active_hashcash/application_controller.rb
@@ -1,0 +1,4 @@
+module ActiveHashcash
+  class ApplicationController < ActionController::Base
+  end
+end