+++ app/helpers/active_hashcash/application_helper.rb
@@ -1,0 +1,4 @@
+module ActiveHashcash
+  module ApplicationHelper
+  end
+end