+++ app/jobs/active_hashcash/application_job.rb
@@ -1,0 +1,4 @@
+module ActiveHashcash
+  class ApplicationJob < ActiveJob::Base
+  end
+end