+++ app/models/active_hashcash/application_record.rb
@@ -1,0 +1,5 @@
+module ActiveHashcash
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end