+++ config/routes.rb
@@ -1,0 +1,2 @@
+ActiveHashcash::Engine.routes.draw do
+end