+++ lib/active_hashcash.rb
@@ -1,3 +1,6 @@
+require "active_hashcash/version"
+require "active_hashcash/engine"
+
 module ActiveHashcash
   extend ActiveSupport::Concern
 
@@ -7,24 +10,16 @@ module ActiveHashcash
     helper_method :hashcash_hidden_field_tag
   end
 
-  mattr_accessor :bits, instance_accessor: false, default: 20
   mattr_accessor :resource, instance_accessor: false
-  mattr_accessor :redis_url, instance_accessor: false, default: ENV["ACTIVE_HASHCASH_REDIS_URL"] || ENV["REDIS_URL"]
-
-  def self.store
-    @store ||= Store.new(Redis.new(url: ActiveHashcash.redis_url))
-  end
-
-  def self.store=(store)
-    @store = store
-  end
+  mattr_accessor :bits, instance_accessor: false, default: 20
+  mattr_accessor :date_format, instance_accessor: false, default: "%y%m%d"
 
   # TODO: protect_from_brute_force bits: 20, exception: ActionController::InvalidAuthenticityToken, with: :handle_failed_hashcash
 
   # Call me via a before_action when the form is submitted : `before_action :check_hashcash, only: :create`
   def check_hashcash
-    stamp = hashcash_param && Stamp.parse(hashcash_param)
-    if stamp && stamp.verify(hashcash_resource, hashcash_bits, Date.yesterday) && ActiveHashcash.store.add?(stamp)
+    #stamp = hashcash_param && Stamp.spend(hashcash_param)
+    if hashcash_param && Stamp.spend(hashcash_param, hashcash_resource, hashcash_bits, Date.yesterday, request.remote_ip)
       hashcash_after_success
     else
       hashcash_after_failure
@@ -72,7 +67,3 @@ module ActiveHashcash
     hidden_field_tag(name, "", "data-hashcash" => options.to_json)
   end
 end
-
-require "active_hashcash/stamp"
-require "active_hashcash/store"
-require "active_hashcash/engine"