+++ test/dummy/config/routes.rb
@@ -1,0 +1,3 @@
+Rails.application.routes.draw do
+  mount ActiveHashcash::Engine => "/active_hashcash"
+end